refact: about section in settings
